Added python bindings for etherbone.h
This is just a quick-and-dirty python layer over the C99
API. It is UNTESTED, only imports correctly, and must
be overriden by a properly written, pythonic set of classes
calling these as methods.

Please note that, accordingly, a new libetherbone.so target
has been added to the Makefile, for ctypes dynamic loading.

This is quick. This is not good. Just for the sake of quick tests, in
case Tom needs it for NOW :D.